Hieu

+++ backend/db/migrate/20221010154809_create_follows.rb
@@ -1,8 +1,8 @@
 class CreateFollows < ActiveRecord::Migration[6.0]
   def change
     create_table :follows do |t|
-      t.bigint :follower_id
-      t.bigint :following_id
+      t.integer :follower_id
+      t.integer :following_id
 
       t.timestamps
     end
